+++ src/rmcs_analyzer/export/__init__.py
@@ -1,5 +1,6 @@
 """Export services for RMCS Analyzer."""
 
+from .burnsim import export_burnsim_csv
 from .pdf_report import (
     PDFReportError,
     generate_campaign_pdf_report,
@@ -10,4 +11,5 @@ __all__ = [
     "PDFReportError",
     "generate_campaign_pdf_report",
     "generate_pdf_report",
+    "export_burnsim_csv",
 ]